4.21: blog deletion uses token authentication

+++ part4/blog-list/controllers/blogs.js
@@ -35,8 +35,17 @@ blogsRouter.post("/", async (request, response) => {
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({error: "token missing or invalid"})
+    }
+    if (decodedToken.id.toString() === blog.user.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        return response.status(204).end()
+    }
+    return response.status(401).json({error: "token does not match with blog author"})
+    
 })
 
 blogsRouter.put("/:id", async (request, response) => {
